% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop_gen.R
\name{sim_pop_gen}
\alias{sim_pop_gen}
\title{Simulate allele and genotype frequencies at a single locus}
\usage{
sim_pop_gen(A = 0.5, a = 0.5, AA = 0, Aa = 0, aa = 0,
  generations = 150, PopSize = 100, w_AA = 1, w_Aa = 1, w_aa = 1,
  M_AA = 0, M_Aa = 0, M_aa = 0, plot_alleles = TRUE,
  plot_genotypes = FALSE, print_data = FALSE, new_plot = TRUE)
}
\arguments{
\item{A}{The initial frequency of the 'A' allele}

\item{a}{The initial frequency of the 'a' allele}

\item{AA}{The initial number of individuals of the 'AA' genotype}

\item{Aa}{The initial number of individuals of the 'Aa' genotype}

\item{aa}{The initial number of individuals of the 'aa' genotype}

\item{generations}{The number of simulations for which the simulation will run}

\item{PopSize}{The size of the population being simulated (this will remain constant over generations)}

\item{w_AA}{The relative fitness of the 'AA' genotype}

\item{w_Aa}{The relative fitness of the 'Aa' genotype}

\item{w_aa}{The relative fitness of the 'aa' genotype}

\item{M_AA}{The number of individuals of genotype 'AA' migrating into the population each generation}

\item{M_Aa}{The number of individuals of genotype 'Aa' migrating into the population each generation}

\item{M_aa}{The number of individuals of genotype 'aa' migrating into the population each generation}

\item{plot_alleles}{Whether or not allele frequencies should be plotted in each generation}

\item{plot_genotypes}{Whether or not genotype frequencies should be plotted in each generation}

\item{print_data}{Whether or not the allele and genotype frequencies should be returned in the form of a table as output}

\item{new_plot}{Whether a new window should be opened to plot}
}
\value{
A plot showing allele frequencies, genotype frequencies, or both, plotted over generations. Output can also include data on allele frequencies and genotype frequencies in each generation.
}
\description{
This function simulates population genetics at a single locus by modelling the change in allele frequencies over generations as a consequence of genetic drift, migration, and natural selection. Arguments allow users to specify starting allele or genotype frequencies, and different genotype fitnesses or population sizes.
}
\examples{
sim_pop_gen(A = 0.7, a = 0.3);
}
